fix: keep choices in sync with available nodes (#4)
Previously, if a choice was added, then later nodes were added/removed,
the available target node options for the choices would not be updated.
This resulted in inability to select nodes,
or choices pointing to removed nodes.
This fixes that by updating all the choice select elements
whenever a node is added or removed.

Also fix bugs where .trim() can be called on undefined values.
Fix them by defaulting to an empty string if the values are undefined.

Private-ref: https://tasks.opencraft.com/browse/BB-9970

Co-authored-by: Pooja Kulkarni <pooja@opencraft.com>